+++ bootstrapgenerator/src/test/java/Utilities.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2019 Open Web IT B.V. (https://www.openweb.nl/)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import java.io.File;
+import java.io.IOException;
+
+import org.codehaus.plexus.util.FileUtils;
+
+public class Utilities {
+    public static void enforceWindowsFileEndings(File file) throws IOException {
+        String content = FileUtils.fileRead(file);
+        String lfContent = content.replaceAll("\r\n", "\n");
+        FileUtils.fileWrite(file, lfContent.replaceAll("\n", "\r\n"));
+    }
+}
